backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
openai
google-generativeai
stripe
firebase-admin
pydantic
